Moved createImageRegionPredicate to PathObjectTools
This avoids direct JTS dependency in IJExtension.
Also reduce lag in syntax highlighting for the script editor.
